requirements.txt
fastapi
uvicorn
sqlalchemy
mysqlclient
python-dotenv
pywebview
